README.md
This script provides automation of scanning in Symantec, AbuseIP and VirusTotal. In the "direcciones_ip.txt" tab you must enter a URL or IP Address per line for future reading of the Script and its operation.
It is a way to streamline the tasks of a SOC Analyst, stopping wasting time scanning different domains that could have generated an alert.
You can copy the Script and use it in the tool of your choice. In my case I use VSCode to save and modify it.
This Script uses the Selenium library for better reading of each of the pages.

You can add more pages in the same script following the specific rules of each line.

REMEMBER TO INSTALL THE COMPLETE LIBRARY OF SELENIUM.